Updated Game's Logic
Actually, the true game's goal consists of making the highest points value choosing to keep a dice which value is added to the partial points obtained until then. When you choose to keep a dice, the dice itself is removed from the set of available dices to be kept in the following rolls. At the end, the result is obtained by summing the values of all the kept dices.

At any roll, you can choose not to keep any dice if the values are too low: in this case, the game continues with the following roll, and it is still possible choosing to keep a dice from the same set of dices available until then. When the player reaches the 3rd roll, all the dices values are summed together and added to his points so far, since he can't choose to keep no more dices.